Update structure from TOneV2_Dev
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C41643

+++ TOneV2/Code/TOneV2/SQL.TOneWhS_BE/TOneWhS_BE/Tables/SalePricelistRPChange_New.sql
@@ -6,7 +6,9 @@
     [BED]                    DATETIME       NULL,
     [EED]                    DATETIME       NULL,
     [PriceListId]            BIGINT         NULL,
-    [CountryId]              BIGINT         NULL,
+    [CountryId]              INT            NULL,
     [ProcessInstanceID]      INT            NULL
 );
+
+
